move networks from v2/config to v2/services/EthService
