+++ common/v2/services/EthService/index.ts
@@ -63,6 +63,7 @@ export {
   isValidGetNetVersion,
   isValidAddress
 } from './validators';
+export { ProviderHandler, ENSProvider } from './network';
 // @TODO These are consummed by v2/libs
 // remove export after migration into the service
 export {
